fix: icon position fix
